queryquest_django_app/requirements.txt
Django==3.1.7
PyMySQL==1.0.2
djangorestframework>=3.14.0